feat(test): enable agent develop mode for local integration tests
